fix hashtags

+++ socialmedia_project/dashboard/models.py
@@ -33,7 +33,7 @@ class Tweet(models.Model):
     text = models.TextField()
     retweets = models.IntegerField(default=0)
     tweeted_at = models.DateTimeField()
-    tweet_id = models.IntegerField(unique=True)
+    tweet_id = models.BigIntegerField(unique=True)
     sentiment = models.TextField()
     by = models.ForeignKey(TwitterFollower)
     spec = models.ForeignKey(UserSpec)
